update video cutter

+++ src/main/java/com/example/youtubeDownloader/controller/YoutubeController.java
@@ -3,6 +3,7 @@ package com.example.youtubeDownloader.controller;
 import com.example.youtubeDownloader.service.YoutubeDownloaderService;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import org.bytedeco.javacpp.indexer.DoubleIndexer;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
@@ -14,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
@@ -27,8 +27,8 @@ public class YoutubeController {
     @SneakyThrows
     @GetMapping("/downloads/{videoId}")
     public ResponseEntity<Resource> downloadVideo(@PathVariable("videoId") String videoId,
-                                                  @RequestParam(value = "startTime", required = false) Long startTime,
-                                                  @RequestParam(value = "endTime", required = false) Long endTime) {
+                                                  @RequestParam(value = "startTime", required = false) Double startTime,
+                                                  @RequestParam(value = "endTime", required = false) Double endTime) {
 
         String outputPath = "downloads/" + videoId;
         File file = youtubeDownloaderService.downloadVideo(videoId, outputPath, startTime, endTime);
